requirements.txt
flask
flask-cors
torch
openai-whisper
sentence-transformers
ytmusicapi
pandas
numpy
ffmpeg-python
